Make the sensor menu setting persistent

+++ rootdir/etc/init.qcom.sh
@@ -293,7 +293,11 @@ case "$target" in
         platformvalue=`cat /sys/devices/system/soc/soc0/hw_platform`
         case "$platformvalue" in
             "Fluid")
-                echo 1 > /data/system/sensors/settings
+                if [ ! -s /data/system/sensors/settings ]; then
+                    # If the settings file is empty, enable sensors
+                    # Otherwise leave the file with it's current contents
+                    echo 1 > /data/system/sensors/settings
+                fi
                 start sensors
                 setprop ro.sf.lcd_density 240
                 start profiler_daemon;;
@@ -308,10 +312,14 @@ case "$target" in
         chmod 220 /sys/devices/platform/msm_hsusb/gadget/wakeup
         ;;
     "msm8960")
-	echo 1 > /data/system/sensors/settings
-	start sensors
-	chown root.system /sys/devices/platform/msm_otg/msm_hsusb/gadget/wakeup
-	chmod 220 /sys/devices/platform/msm_otg/msm_hsusb/gadget/wakeup
+        if [ ! -s /data/system/sensors/settings ]; then
+            # If the settings file is empty, enable sensors
+            # Otherwise leave the file with it's current contents
+            echo 1 > /data/system/sensors/settings
+        fi
+        start sensors
+        chown root.system /sys/devices/platform/msm_otg/msm_hsusb/gadget/wakeup
+        chmod 220 /sys/devices/platform/msm_otg/msm_hsusb/gadget/wakeup
         ;;
     "msm7630_surf" )
         chown root.system /sys/devices/platform/msm_hsusb/gadget/wakeup
